separate ExitUsecase
